[DELETE] class conflicts
