+++ src/main/java/com/dsmbamboo/api/configs/SecurityConfiguration.java
@@ -3,10 +3,13 @@ package com.dsmbamboo.api.configs;
 import com.dsmbamboo.api.domains.users.filter.JwtAuthenticationFilter;
 import com.dsmbamboo.api.domains.users.filter.JwtAuthorizationFilter;
 import com.dsmbamboo.api.domains.users.model.UserRepository;
-import com.dsmbamboo.api.domains.users.service.UserPrincipalDetailsService;
+import com.dsmbamboo.api.domains.users.security.JwtConfigurer;
+import com.dsmbamboo.api.domains.users.security.JwtTokenProvider;
+import com.dsmbamboo.api.domains.users.security.UserPrincipalDetailsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -23,6 +26,14 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 
     private final UserPrincipalDetailsService userPrincipalDetailsService;
     private final UserRepository userRepository;
+
+    private final JwtTokenProvider jwtTokenProvider;
+
+    @Bean
+    @Override
+    public AuthenticationManager authenticationManagerBean() throws Exception {
+        return super.authenticationManagerBean();
+    }
 
     @Override
     protected void configure(AuthenticationManagerBuilder auth) {
@@ -47,7 +58,10 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
                     .addFilter(new JwtAuthenticationFilter(authenticationManager()))
                     .addFilter(new JwtAuthorizationFilter(authenticationManager(), userRepository))
                 .authorizeRequests()
-                    .anyRequest().permitAll();
+                    .antMatchers("/auth/signin").permitAll()
+                    .anyRequest().permitAll()
+                .and()
+                    .apply(new JwtConfigurer(jwtTokenProvider));
 
     }
