+++ src/main/java/com/dsmbamboo/api/domains/users/security/UserPrincipalDetailsService.java
@@ -1,24 +1,23 @@
 package com.dsmbamboo.api.domains.users.security;
 
 import com.dsmbamboo.api.domains.users.model.UserRepository;
+import lombok.AllArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
+@AllArgsConstructor
 public class UserPrincipalDetailsService implements UserDetailsService {
 
     private UserRepository userRepository;
-
-    public UserPrincipalDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return userRepository.findByUsername(username)
                 .map(UserPrincipal::new)
-                .orElseThrow(() -> new UsernameNotFoundException("Username:" + username + "not found."));
+                .orElseThrow(() -> new UsernameNotFoundException("username has not found."));
     }
+
 }
