--- src/main/java/com/dsmbamboo/api/domains/users/service/UserPrincipalDetailsService.java
@@ -1,24 +1,0 @@
-package com.dsmbamboo.api.domains.users.service;
-
-import com.dsmbamboo.api.domains.users.model.User;
-import com.dsmbamboo.api.domains.users.model.UserRepository;
-import com.dsmbamboo.api.domains.users.security.UserPrincipal;
-import lombok.AllArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class UserPrincipalDetailsService implements UserDetailsService {
-
-    private UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username);
-        return new UserPrincipal(user);
-    }
-
-}
